pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "shared-ledger-system"
version = "0.1.0"
description = "A shared ledger system for multiple applications in a monorepo."
authors = [
    { name = "Tarık Buğra Çat", email = "tarikbugra.cat@gmail.com" }
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.95.0",
    "sqlalchemy>=2.0.0",
    "pydantic>=1.10.0",
    "alembic>=1.10.0",
    "psycopg2-binary>=2.9.0",
    "uvicorn[standard]>=0.21.0",
    "python-dotenv>=1.0.1",
    "enum-tools>=0.10.0",
]

[project.optional-dependencies]
dev = [
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "sqlalchemy2-stubs>=0.0.2a20",
]

[tool.setuptools.packages.find]
include = ["shared_ledger_system*"]

[tool.black]
line-length = 88
target-version = ['py310']

[tool.flake8]
max-line-length = 88
ignore = "E203, E266, E501, W503"
exclude = [".venv", "venv", "migrations", "tests"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-v --cov=."
testpaths = ["tests"]

[tool.alembic]
script_location = "migrations"
sqlalchemy.url = "postgresql+psycopg2://${DB_USER}:${DB_PASS}@${DB_HOST}/${DB_NAME}"